+++ core/lib/config/src/configs/node_sync.rs
@@ -12,6 +12,9 @@ pub struct NodeSyncConfig {
     /// Maximum number of transactions to process in a single batch
     #[config(default_t = NonZeroU64::new(10_000).unwrap())]
     pub batch_transaction_updater_batch_size: NonZeroU64,
+    /// Toggle for disabling seal criteria validation in case of some issues / forced proved batches
+    #[config(default_t = true)]
+    pub validate_seal_criteria: bool,
 }
 
 #[cfg(test)]
@@ -24,6 +27,7 @@ mod tests {
         NodeSyncConfig {
             batch_transaction_updater_interval: Duration::from_secs(2),
             batch_transaction_updater_batch_size: NonZeroU64::new(100).unwrap(),
+            validate_seal_criteria: false,
         }
     }
 
@@ -32,6 +36,7 @@ mod tests {
         let env = r#"
             NODE_SYNC_BATCH_TRANSACTION_UPDATER_INTERVAL=2sec
             NODE_SYNC_BATCH_TRANSACTION_UPDATER_BATCH_SIZE=100
+            NODE_SYNC_VALIDATE_SEAL_CRITERIA=false
         "#;
         let env = Environment::from_dotenv("test.env", env)
             .unwrap()
@@ -46,6 +51,7 @@ mod tests {
         let yaml = r#"
           batch_transaction_updater_interval: 2sec
           batch_transaction_updater_batch_size: 100
+          validate_seal_criteria: false
         "#;
         let yaml = Yaml::new("test.yml", serde_yaml::from_str(yaml).unwrap()).unwrap();
         let config: NodeSyncConfig = test_complete(yaml).unwrap();