+++ zkstack_cli/crates/zkstack/src/commands/chain/genesis/mod.rs
@@ -1,7 +1,7 @@
 use clap::{command, Parser, Subcommand};
 use xshell::Shell;
 use zkstack_cli_common::{logger, spinner::Spinner};
-use zkstack_cli_config::{ChainConfig, ZkStackConfig};
+use zkstack_cli_config::{ChainConfig, ZkStackConfig, ZkStackConfigTrait};
 
 use crate::{
     commands::chain::{
@@ -48,19 +48,19 @@ pub async fn run_genesis(args: GenesisArgs, shell: &Shell) -> anyhow::Result<()>
     let chain_config = ZkStackConfig::current_chain(shell)?;
     let args = args.fill_values_with_prompt(&chain_config);
 
-    genesis(args, shell, &chain_config).await?;
+    genesis(&args, shell, &chain_config).await?;
     logger::outro(MSG_GENESIS_COMPLETED);
 
     Ok(())
 }
 
 pub async fn genesis(
-    args: GenesisArgsFinal,
+    args: &GenesisArgsFinal,
     shell: &Shell,
     config: &ChainConfig,
 ) -> anyhow::Result<()> {
     let override_validium_config = true;
-    database::update_configs(args.clone(), shell, config, override_validium_config).await?;
+    database::update_configs(args, shell, config, override_validium_config).await?;
 
     logger::note(
         MSG_SELECTED_CONFIG,
@@ -72,12 +72,17 @@ pub async fn genesis(
     logger::info(MSG_STARTING_GENESIS);
 
     let spinner = Spinner::new(MSG_INITIALIZING_DATABASES_SPINNER);
-    initialize_server_database(shell, &args.server_db, &config.link_to_code, args.dont_drop)
-        .await?;
+    initialize_server_database(
+        shell,
+        &args.server_db,
+        &config.link_to_code(),
+        args.dont_drop,
+    )
+    .await?;
     spinner.finish();
 
     let spinner = Spinner::new(MSG_STARTING_GENESIS_SPINNER);
-    run_server_genesis(args.server_command, config, shell)?;
+    run_server_genesis(args.server_command.clone(), config, shell)?;
     spinner.finish();
 
     Ok(())